feat: switch to SSG
