+++ examples/gatsby5/web/gatsby-browser.tsx
@@ -1,26 +1,34 @@
+/**
+ * Implement Gatsby's Browser APIs in this file.
+ *
+ * See: https://www.gatsbyjs.com/docs/browser-apis/
+ */
+// You can delete this file if you're not using it
 import "./src/styles/global.css";
-import { UniformContext } from "@uniformdev/context-react";
-import {
-  Context,
-  enableContextDevTools,
-  ManifestV2,
-} from "@uniformdev/context";
-import manifest from "./uniform-manifest.json";
-import { GatsbyBrowser } from "gatsby";
-import * as React from "react";
 
-const context = new Context({
-  defaultConsent: true,
-  plugins: [enableContextDevTools()], // Use this to enable the Chrome plugin
-  manifest: manifest as ManifestV2,
-});
+// import "./src/styles/global.css";
+// import { UniformContext } from "@uniformdev/context-react";
+// import {
+//   Context,
+//   enableContextDevTools,
+//   ManifestV2,
+// } from "@uniformdev/context";
+// import manifest from "./uniform-manifest.json";
+// import { GatsbyBrowser } from "gatsby";
+// import * as React from "react";
 
-export const wrapPageElement: GatsbyBrowser["wrapPageElement"] = ({
-  element,
-}) => {
-  return (
-    <UniformContext context={context} outputType={"standard"}>
-      {element}
-    </UniformContext>
-  );
-};
+// const context = new Context({
+//   defaultConsent: true,
+//   plugins: [enableContextDevTools()], // Use this to enable the Chrome plugin
+//   manifest: manifest as ManifestV2,
+// });
+
+// export const wrapPageElement: GatsbyBrowser["wrapPageElement"] = ({
+//   element,
+// }) => {
+//   return (
+//     <UniformContext context={context} outputType={"standard"}>
+//       {element}
+//     </UniformContext>
+//   );
+// };
