+++ examples/gatsby5/web/src/components/Container.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Layout } from "./Layout";
 
-const PageComponent = (props: any) => {
+const Container = (props: any) => {
   return (
     <Layout>
       <div className="container mx-auto">{props.children}</div>
@@ -9,4 +9,4 @@ const PageComponent = (props: any) => {
   );
 };
 
-export default PageComponent;
+export default Container;
