+++ examples/gatsby5/web/src/pages/main.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Home() {
+  return (
+    <h1 className="text-xxl">
+      Dummy page, please ignore. Index page will be generated dynamically.
+    </h1>
+  );
+}
